+++ scripts/utils/settings.js
@@ -195,7 +195,7 @@ export function registerSettings() {
     scope: 'client',
     config: false,
     type: Boolean,
-    default: false,
+    default: true,
     onChange: () => ui.BG3HUD_APP?.updateDisplaySettings()
   });